EL-2885 Make the login components available and add documentation Move notifications-wrapper.directive.js to notifications-wrapper.directive.ts.
